+++ src/Item/DefaultItem.php
@@ -11,6 +11,9 @@ namespace HeimrichHannot\ReaderBundle\Item;
 use Contao\DataContainer;
 use Contao\StringUtil;
 use Contao\System;
+use HeimrichHannot\ConfigElementTypeBundle\ConfigElementType\ConfigElementData;
+use HeimrichHannot\ConfigElementTypeBundle\ConfigElementType\ConfigElementResult;
+use HeimrichHannot\ConfigElementTypeBundle\ConfigElementType\ConfigElementTypeInterface;
 use HeimrichHannot\ReaderBundle\ConfigElementType\ConfigElementType;
 use HeimrichHannot\ReaderBundle\ConfigElementType\ReaderConfigElementData;
 use HeimrichHannot\ReaderBundle\Event\ReaderAfterRenderEvent;
@@ -273,7 +276,23 @@ class DefaultItem implements ItemInterface, \JsonSerializable
         if (null !== ($readerConfigElements = $this->_manager->getReaderConfigElementRegistry()->findBy(['tl_reader_config_element.pid=?'], [$readerConfig->rootId]))) {
             foreach ($readerConfigElements as $readerConfigElement) {
                 if ($readerConfigElementType = $this->_manager->getReaderConfigElementRegistry()->getReaderConfigElementType($readerConfigElement->type)) {
-                    $readerConfigElementType->addToReaderItemData(new ReaderConfigElementData($this, $readerConfigElement));
+                    if ($readerConfigElementType instanceof ConfigElementTypeInterface) {
+                        $result = $readerConfigElementType->applyConfiguration(new ConfigElementData($this->getRaw(), $readerConfigElement));
+
+                        switch ($result->getType()) {
+                            case ConfigElementResult::TYPE_FORMATTED_VALUE:
+                                $this->setFormattedValue($readerConfigElement->templateVariable, $result->getValue());
+
+                                break;
+
+                            case ConfigElementResult::TYPE_RAW_VALUE:
+                                $this->setRawValue($readerConfigElement->templateVariable, $result->getValue());
+
+                                break;
+                        }
+                    } else {
+                        $readerConfigElementType->addToReaderItemData(new ReaderConfigElementData($this, $readerConfigElement));
+                    }
                 } else {
                     if (null === ($class = $this->_manager->getReaderConfigElementRegistry()->getElementClassByName($readerConfigElement->type))) {
                         continue;