+++ src/Resources/contao/config/config.php
@@ -1,14 +1,16 @@
 <?php
 
-/**
- * Backend modules
+/*
+ * Copyright (c) 2020 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
  */
 
 $GLOBALS['BE_MOD']['system']['reader_configs'] = [
-    'tables' => ['tl_reader_config']
+    'tables' => ['tl_reader_config'],
 ];
 
-/**
+/*
  * Frontend modules
  */
 array_insert(
@@ -19,14 +21,14 @@ array_insert(
     ]
 );
 
-/**
+/*
  * Permissions
  */
 $GLOBALS['TL_PERMISSIONS'][] = 'readerbundles';
 $GLOBALS['TL_PERMISSIONS'][] = 'readerbundlep';
 
-/**
+/*
  * Models
  */
-$GLOBALS['TL_MODELS']['tl_reader_config']         = \HeimrichHannot\ReaderBundle\Model\ReaderConfigModel::class;
+$GLOBALS['TL_MODELS']['tl_reader_config'] = \HeimrichHannot\ReaderBundle\Model\ReaderConfigModel::class;
 $GLOBALS['TL_MODELS']['tl_reader_config_element'] = \HeimrichHannot\ReaderBundle\Model\ReaderConfigElementModel::class;